enable winsigner to use the new signing cert

+++ buildSrc/winsigner.cjs
@@ -1,6 +1,6 @@
-const path = require('path')
+const path = require("path")
 const fs = require("fs-extra")
-const spawn = require('child_process').spawn
+const spawn = require("child_process").spawn
 
 const TAG = "WINSIGN"
 
@@ -12,16 +12,16 @@ const TAG = "WINSIGN"
  * argument names may be fixed by electron-builder
  */
 function signer({
-					path: pathToSign, // path to the file to sign (string)
-					hash: hashAlgorithm // hash algorithm to use (string, defaults to "sha256")
-				}) {
+	path: pathToSign, // path to the file to sign (string)
+	hash: hashAlgorithm, // hash algorithm to use (string, defaults to "sha256")
+}) {
 	const ext = path.extname(pathToSign)
 	// /thing/thong.AppImage -> /thing/thong-unsigned.AppImage
 	const unsignedFileName = pathToSign.slice(0, pathToSign.length - ext.length) + "-unsigned" + ext
 	console.log(TAG, `signing ${unsignedFileName} as ${pathToSign}`)
 	const commandArguments = process.env.DEBUG_SIGN
 		? getSelfSignedArgs(unsignedFileName, hashAlgorithm, pathToSign)
-		: getHsmArgs(unsignedFileName, hashAlgorithm, pathToSign)
+		: getTokenArgs(unsignedFileName, hashAlgorithm, pathToSign)
 
 	return signWithArgs(commandArguments, pathToSign, unsignedFileName)
 }
@@ -31,18 +31,23 @@ function getSelfSignedArgs(unsignedFileName, hash, signedFileOutPath) {
 
 	return [
 		"sign",
-		"-in", unsignedFileName,
-		"-out", signedFileOutPath,
-		"-pkcs12", certificateFile,
-		"-h", hash ? hash : "sha256",
-		"-t", "http://timestamp.comodoca.com/authenticode",
-		"-n", "tutanota-desktop"
+		"-in",
+		unsignedFileName,
+		"-out",
+		signedFileOutPath,
+		"-pkcs12",
+		certificateFile,
+		"-h",
+		hash ? hash : "sha256",
+		"-t",
+		"http://timestamp.comodoca.com/authenticode",
+		"-n",
+		"tutanota-desktop",
 	]
 }
 
-function getHsmArgs(unsignedFileName, hash, signedFileOutPath) {
-	const certificateFile = process.env["WIN_CSC_FILE"]
-	const hsmPin = process.env["HSM_USER_PIN"]
+function getTokenArgs(unsignedFileName, hash, signedFileOutPath) {
+	const hsmPin = process.env["YUBI_PIN"]
 
 	//  Timestamping:
 	//  1. The client application creates a hashed value of the data to the timestamp server.
@@ -56,31 +61,36 @@ function getHsmArgs(unsignedFileName, hash, signedFileOutPath) {
 	//  http://www.startssl.com/timestamp
 	//  http://timestamp.sectigo.com
 
-	if (!certificateFile) {
-		console.error("ERROR: " + signedFileOutPath.split(path.sep).pop() + "\" not signed! The NSIS installer may not work.")
-		console.log("\t• set WIN_CSC_FILE env var")
-		throw new Error(signedFileOutPath)
-	}
-
 	if (!hsmPin) {
-		console.log("ERROR: " + signedFileOutPath.split(path.sep).pop() + "\" not signed! The NSIS installer may not work.")
+		console.log("ERROR: " + signedFileOutPath.split(path.sep).pop() + '" not signed! The NSIS installer may not work.')
 		console.log("\t• set  HSM_USER_PIN env var")
 		throw new Error(signedFileOutPath)
 	}
 
 	return [
 		"sign",
-		"-in", unsignedFileName,
-		"-out", signedFileOutPath,
+		"-in",
+		unsignedFileName,
+		"-out",
+		signedFileOutPath,
 		// Let the library find the engine, it seems to be more reliable
 		// "-pkcs11engine", "/usr/lib/x86_64-linux-gnu/engines-3/pkcs11.so",
-		"-pkcs11module", "/usr/lib/x86_64-linux-gnu/opensc-pkcs11.so",
-		"-certs", certificateFile,
-		"-key", "11", // this is the key corresponding to the Windows authenticode codesigning certificate
-		"-pass", hsmPin,
-		"-h", hash ? hash : "sha256",
-		"-t", "http://timestamp.comodoca.com/authenticode",
-		"-n", "tutanota-desktop"
+		"-pkcs11module",
+		"/usr/lib/x86_64-linux-gnu/opensc-pkcs11.so",
+		"-pkcs11cert",
+		// serial was acquired by running pkcs11-tool -L whith the yubikey inserted
+		"pkcs11:id=%01;type=cert;serial=6136b562afb4ed9d",
+		"-key",
+		// this is the key corresponding to the Windows authenticode codesigning certificate
+		"pkcs11:id=%01;type=private;serial=6136b562afb4ed9d",
+		"-pass",
+		hsmPin,
+		"-h",
+		hash ? hash : "sha256",
+		"-t",
+		"http://timestamp.comodoca.com/authenticode",
+		"-n",
+		"tutanota-desktop",
 	]
 }
 
@@ -92,19 +102,23 @@ function signWithArgs(commandArguments, signedFileOutPath, unsignedFileName, att
 		console.log("\t• install osslsigncode")
 		return Promise.reject(new Error(signedFileOutPath))
 	}
-	console.log(TAG, `renaming ${signedFileOutPath} -> ${unsignedFileName}`)
 	console.log(TAG, "ir:", fs.readdirSync(path.dirname(signedFileOutPath)).join(", "))
-	fs.renameSync(signedFileOutPath, unsignedFileName)
+	if (fs.existsSync(unsignedFileName) && !fs.existsSync(signedFileOutPath)) {
+		console.log(TAG, "skipping rename, probably in retry")
+	} else {
+		console.log(TAG, `renaming ${signedFileOutPath} -> ${unsignedFileName}`)
+		fs.renameSync(signedFileOutPath, unsignedFileName)
+	}
 	console.log(TAG, `running ${command}`)
 	// only for testing, would print certificate password to logs, otherwise
 	//console.log(`spawning "${command} ${commandArguments.join(" ")}"`)
 	const child = spawn(command, commandArguments, {
 		detached: false,
-		stdio: ['ignore', 'inherit', 'inherit'],
+		stdio: ["ignore", "inherit", "inherit"],
 	})
 
 	return new Promise((resolve, reject) => {
-		child.on('close', (exitCode) => {
+		child.on("close", (exitCode) => {
 			if (exitCode === 255 && attempt < 3) {
 				const delayMs = (attempt + 1) * 10 * 1000
 				console.log(TAG, `signing failed with 255 for ${unsignedFileName}, trying out again with delay of ${delayMs}ms`)
